+++ crates/core/src/database/drivers/sql/create_ipbans.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS ipbans (
     ip TEXT NOT NULL,
-    created INTEGER NOT NULL PRIMARY KEY,
+    created BIGINT NOT NULL PRIMARY KEY,
     reason TEXT NOT NULL,
     moderator TEXT NOT NULL
 )
