+++ crates/core/src/database/drivers/sql/create_memberships.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS memberships (
-    id INTEGER NOT NULL PRIMARY KEY,
-    created INTEGER NOT NULL,
-    owner INTEGER NOT NULL,
-    community INTEGER NOT NULL,
-    role INTEGER NOT NULL
+    id BIGINT NOT NULL PRIMARY KEY,
+    created BIGINT NOT NULL,
+    owner BIGINT NOT NULL,
+    community BIGINT NOT NULL,
+    role BIGINT NOT NULL
 )
