+++ crates/core/src/database/drivers/sql/create_reports.sql
@@ -1,8 +1,8 @@
 CREATE TABLE IF NOT EXISTS reports (
-    id INTEGER NOT NULL PRIMARY KEY,
-    created INTEGER NOT NULL,
-    owner INTEGER NOT NULL,
+    id BIGINT NOT NULL PRIMARY KEY,
+    created BIGINT NOT NULL,
+    owner BIGINT NOT NULL,
     content TEXT NOT NULL,
-    asset INTEGER NOT NULL,
+    asset BIGINT NOT NULL,
     asset_type TEXT NOT NULL
 )
